Use int instead of uint in some more places

+++ num_test.go
@@ -836,13 +836,13 @@ func TestCoprimeExamples(t *testing.T) {
 
 func TestTrueLenExamples(t *testing.T) {
 	x := new(Nat).SetUint64(0x0000_0000_0000_0001)
-	expected := uint(1)
+	expected := 1
 	actual := x.TrueLen()
 	if expected != actual {
 		t.Errorf("%+v != %+v", expected, actual)
 	}
 	x.SetUint64(0x0000_0000_0100_0001)
-	expected = uint(25)
+	expected = 25
 	actual = x.TrueLen()
 	if expected != actual {
 		t.Errorf("%+v != %+v", expected, actual)
